Higher quantity and fix issue #28

+++ database/migrations/2015_05_17_115707_create_ingredients.php
@@ -18,7 +18,7 @@ class CreateIngredients extends Migration {
 		{
 			$table->increments('id');
             $table->char('ingredient',50);
-            $table->tinyInteger('stock')->unsigned();
+            $table->smallInteger('stock')->unsigned();
             $table->tinyInteger('position')->nullable();
 		});
 	}
